Fix Command Race Condition (#11)
Because the `GitHubActionsCommands` class was writing directly to the
console, it meant that messages were arriving in the logs out of order,
because `ILogger` uses buffering.

This fixes the issue by updating the command runner to go through
`ILogger` instead, using a well known `EventId` as a way to instruct the
log formatter not to add its usual formatting.

Additionally, I added a new `WithGroup` method to the command runner
that lets you dispose an object to close the group just to make using it
a bit nicer.